# docs/contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bookstore Microservices API",
    "description": "Complete API specification for the Bookstore microservices platform including Catalog, Order, and User services accessible through the API Gateway",
    "version": "1.0.0",
    "contact": {
      "name": "Bookstore API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "API Gateway - Local Development"
    }
  ],
...
  "paths": {
...
    "/api/v1/orders/user/{userId}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get user orders",
        "description": "Retrieves all orders for a specific user",
        "operationId": "getUserOrders",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "ID of the user",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User orders retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/profile": {
      "get": {
        "tags": ["Users"],
        "summary": "Get current user profile",
        "description": "Returns the authenticated user's profile",
        "operationId": "getCurrentUserProfile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update current user profile",
        "description": "Updates the authenticated user's profile",
        "operationId": "updateCurrentUserProfile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by ID",
        "description": "Returns user profile by ID (Admin only)",
        "operationId": "getUserById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to retrieve",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user",
        "description": "Deletes a user account (Admin only)",
        "operationId": "deleteUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to delete",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication token. Include in Authorization header as 'Bearer {token}'"
      }
    },
    "schemas": {
...
      "OrderDto": {
        "type": "object",
        "required": ["userId", "items"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the order (MongoDB ObjectId)",
            "readOnly": true,
            "example": "507f1f77bcf86cd799439011"
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "description": "Identifier of the user who placed the order",
            "example": 1
          },
          "items": {
            "type": "array",
            "minItems": 1,
            "description": "List of items in the order",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          },
          "totalAmount": {
            "type": "number",
            "format": "decimal",
            "description": "Total amount for the entire order (auto-calculated)",
            "readOnly": true,
            "example": 45.97
          },
          "status": {
            "type": "string",
            "enum": ["PENDING", "CONFIRMED", "SHIPPED", "DELIVERED", "CANCELLED"],
            "description": "Current status of the order",
            "example": "PENDING"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the order was created",
            "readOnly": true,
            "example": "2026-01-05T21:13:00Z"
          }
        }
      },
      "OrderItemDto": {
        "type": "object",
        "required": ["bookId", "quantity"],
        "properties": {
          "bookId": {
            "type": "integer",
            "format": "int64",
            "description": "Identifier of the book being ordered",
            "example": 1
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "description": "Quantity of books ordered",
            "example": 2
          },
          "price": {
            "type": "number",
            "format": "decimal",
            "description": "Price per unit at the time of order (auto-populated)",
            "readOnly": true,
            "example": 12.99
          },
          "bookTitle": {
            "type": "string",
            "description": "Title of the book (captured at order time, auto-populated)",
            "readOnly": true,
            "example": "The Great Gatsby"
          },
          "bookIsbn": {
            "type": "string",
            "description": "ISBN of the book (captured at order time, auto-populated)",
            "readOnly": true,
            "example": "9781234567890"
          }
        }
      },
      "UserDto": {
        "type": "object",
        "required": ["email", "firstName", "lastName", "role"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique identifier for the user",
            "readOnly": true,
            "example": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 255,
            "description": "Email address of the user (must be unique)",
            "example": "user@example.com"
          },
          "firstName": {
            "type": "string",
            "maxLength": 100,
            "description": "First name of the user",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "maxLength": 100,
            "description": "Last name of the user",
            "example": "Doe"
          },
          "role": {
            "type": "string",
            "enum": ["CUSTOMER", "ADMIN"],
            "description": "Role of the user in the system",
            "example": "CUSTOMER"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the error occurred",
            "example": "2026-01-05T21:13:00"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code",
            "example": 404
          },
          "error": {
            "type": "string",
            "description": "Short error type description",
            "example": "Not Found"
          },
          "message": {
            "type": "string",
            "description": "Detailed error message explaining what went wrong",
            "example": "Book with ID 999 not found"
          },
          "path": {
            "type": "string",
            "description": "The request path that caused the error",
            "example": "/api/v1/books/999"
          }
        }
      }
    }
  }
}